fix: 抽象permissions-serializer的actions字段值设置
